Separates color options from Color
