+++ src/imgui/imguiWrapper/colorOptions.hpp
@@ -1,0 +1,27 @@
+#ifndef CASH_OVERFLOW_COLOROPTIONS_H
+#define CASH_OVERFLOW_COLOROPTIONS_H
+
+#include "color.hpp"
+
+namespace cash_overflow::colors {
+
+const Color Black{ .0F, .0F, .0F, 1.0F };
+const Color White{ 1.0F, 1.0F, 1.0F, 1.0F };
+const Color Red{ 1.0F, .0F, .0F, 1.0F };
+const Color Lime{ .0F, 1.0F, .0F, 1.0F };
+const Color Blue{ .0F, .0F, 1.0F, 1.0F };
+const Color Yellow{ 1.0F, 1.0F, .0F, 1.0F };
+const Color Cyan{ .1F, 1.0F, 1.0F, 1.0F };
+const Color Magenta{ 1.0F, .0F, 1.0F, 1.0F };
+const Color Silver{ .75F, .75F, .75F, 1.0F };
+const Color Gray{ .5F, .5F, .5F, 1.0F };
+const Color Maroon{ .5F, .0F, .0F, 1.0F };
+const Color Olive{ .5F, .5F, .0F, 1.0F };
+const Color Green{ .0F, .5F, .0F, 1.0F };
+const Color Purple{ .5F, .0F, .5F, 1.0F };
+const Color Teal{ .0F, .5F, .5F, 1.0F };
+const Color Navy{ .0F, .0F, .5F, 1.0F };
+
+}// cash_overflow::colors
+
+#endif// CASH_OVERFLOW_COLOROPTIONS_H
