--- src/imgui/imguiWrapper/colors.hpp
@@ -1,41 +1,0 @@
-#ifndef CASH_OVERFLOW_COLORS_H
-#define CASH_OVERFLOW_COLORS_H
-
-#include <imgui.h>
-
-namespace cash_overflow::colors {
-
-class Color
-{
-public:
-  Color() : color{ 0.0f, 0.0f, 0.0f, 0.0f } {};
-  Color(float r, float g, float b, float o) : color{ r, g, b, o } {};
-  ImVec4 getColor() const
-  {
-    return color;
-  }
-
-private:
-  ImVec4 color{};
-};
-
-  const Color Black{ .0F, .0F, .0F, 1.0F };
-  const Color White{ 1.0F, 1.0F, 1.0F, 1.0F };
-  const Color Red{ 1.0F, .0F, .0F, 1.0F };
-  const Color Lime{ .0F, 1.0F, .0F, 1.0F };
-  const Color Blue{ .0F, .0F, 1.0F, 1.0F };
-  const Color Yellow{ 1.0F, 1.0F, .0F, 1.0F };
-  const Color Cyan{ .1F, 1.0F, 1.0F, 1.0F };
-  const Color Magenta{ 1.0F, .0F, 1.0F, 1.0F };
-  const Color Silver{ .75F, .75F, .75F, 1.0F };
-  const Color Gray{ .5F, .5F, .5F, 1.0F };
-  const Color Maroon{ .5F, .0F, .0F, 1.0F };
-  const Color Olive{ .5F, .5F, .0F, 1.0F };
-  const Color Green{ .0F, .5F, .0F, 1.0F };
-  const Color Purple{ .5F, .0F, .5F, 1.0F };
-  const Color Teal{ .0F, .5F, .5F, 1.0F };
-  const Color Navy{ .0F, .0F, .5F, 1.0F };
-
-}// cash_overflow::colors
-
-#endif//CASH_OVERFLOW_COLORS_H
